添加当接收端列表中没有发送端的项目时候，将发送端添加
