+++ FantasyRemoteCopy.UI/ViewModels/HomePageModel.cs
@@ -167,119 +167,6 @@ public partial class HomePageModel : ViewModelBase, IPageKeep, INavigationAware
     }
 
 
-    private async void CheckPortEnable(TransformResultModel<string> data)
-    {
-        var codeWord = data.Result.ToObject<CodeWordModel>();
-        if (codeWord.Type == CodeWordType.CheckingPort)
-        {
-            var port = codeWord.Port;
-            logger.LogInformation($"端口{port}是否可用");
-            var checkResult = await _portCheckable.IsPortInUse(port);
-            logger.LogInformation($"端口{port}可用状态为 {checkResult}");
-            var sendModel = new SendTextModel(localDevice.Flag, data.Flag,
-                !checkResult
-                    ? CodeWordModel.CreateCodeWord(codeWord.TaskGuid, CodeWordType.CheckingPortCanUse, localDevice.Flag,
-                            data.Flag, port,
-                            codeWord.SendType, localDevice, null)
-                        .ToJson()
-                    : CodeWordModel.CreateCodeWord(codeWord.TaskGuid, CodeWordType.CheckingPortCanNotUse,
-                        localDevice.Flag,
-                        data.Flag,
-                        port,
-                        codeWord.SendType, localDevice, null).ToJson(),
-                ConstParams.TCP_PORT);
-            logger.LogInformation($"端口可用状态信息发送给{data.Flag}");
-            if (!checkResult)
-            {
-                // 开始监听
-                var cancelTokenSource = new CancellationTokenSource();
-                var receiveDevice = DiscoveredDevices.FirstOrDefault(x => x.Flag == data.Flag);
-                if (receiveDevice is null)
-                {
-                    logger.LogInformation($"设备列表中未发现端口为{data.Flag},取消监听");
-                    return;
-                }
-
-                if (receiveDevice.TransmissionTasks.Any(
-                        x => x.TaskGuid == codeWord.TaskGuid))
-                    logger.LogInformation($"{data.Flag}中已经包含了 {data.Flag}-{port} 的任务");
-                else
-                    receiveDevice.TransmissionTasks.Add(new TransmissionTaskModel(codeWord.TaskGuid,
-                        localDevice.Flag, codeWord.Flag, codeWord.Port, codeWord.SendType, cancelTokenSource));
-
-                _tcpLoopListenContentBase.ReceiveAsync(result =>
-                {
-                    // 保存到数据库
-                    SaveDataToLocalDB(result);
-                    if (receiveDevice.TryGetTransmissionTask(codeWord.TaskGuid,
-                            out var v))
-                    {
-                        v?.CancellationTokenSource.Cancel();
-                        receiveDevice.RemoveTransmissionTask(codeWord.TaskGuid);
-                    }
-
-                    SendCommand.NotifyCanExecuteChanged();
-                }, IPAddress.Parse(data.Flag), port, ReportProgress(false, codeWord.TaskGuid), cancelTokenSource.Token);
-            }
-
-            await _tcpSendTextBase.SendAsync(sendModel, null, default);
-        }
-        else if (codeWord.Type == CodeWordType.CancelTransmission)
-        {
-            foreach (var device in DiscoveredDevices)
-            {
-                if (device.TryGetTransmissionTask(codeWord.TaskGuid, out var task))
-                {
-                    task?.CancellationTokenSource?.Cancel();
-                    device.TransmissionTasks.Remove(task);
-                    device.Progress = 0;
-                    device.WorkState = WorkState.None;
-                    break;
-                }
-            }
-
-            SendCommand.NotifyCanExecuteChanged();
-        }
-        else
-        {
-            logger.LogInformation($"发送方接收回调方{data.Flag}端口可用情况，接收方端口可用情况为 {codeWord.Type}");
-
-            //端口不可用，进行累加
-            if (codeWord.Type == CodeWordType.CheckingPortCanNotUse)
-            {
-                var port = codeWord.Port + 1;
-                logger.LogInformation($"接收方{data.Flag}对于{codeWord.Port} 端口无法使用，所以向接收方再次发送{port}端口是否可用");
-                var portCheckMessage = new SendTextModel(localDevice.Flag,
-                    data.Flag ?? throw new NullReferenceException(),
-                    CodeWordModel.CreateCodeWord(codeWord.TaskGuid, CodeWordType.CheckingPort, localDevice.Flag,
-                            data.Flag, port,
-                            codeWord.SendType, localDevice, null)
-                        .ToJson()
-                    , ConstParams.TCP_PORT);
-                await _tcpSendTextBase.SendAsync(portCheckMessage, null, default);
-            }
-            else if (codeWord.Type == CodeWordType.CheckingPortCanUse)
-            {
-                var sendCancelTokenSource = new CancellationTokenSource();
-
-                SenderAddTaskAndShowProgress(codeWord, sendCancelTokenSource);
-
-                switch (InformationModel!.SendType)
-                {
-                    case SendType.Text:
-                        _tcpSendTextBase.SendAsync(
-                            new SendTextModel(localDevice.Flag, data.Flag, InformationModel.Text, codeWord.Port), null,
-                            sendCancelTokenSource.Token);
-                        break;
-                    case SendType.File:
-                    case SendType.Folder:
-                        SendFileAsync(data.Flag, codeWord.Port, sendCancelTokenSource.Token, codeWord.TaskGuid);
-                        break;
-                }
-            }
-        }
-    }
-
     private void SenderAddTaskAndShowProgress(CodeWordModel codeWord, CancellationTokenSource cancelTokenSource)
     {
         var device = DiscoveredDevices.FirstOrDefault(x => x.Flag == codeWord.Flag);
